+++ govee_api_laggat/__init__.py
@@ -1,2 +1,2 @@
-from .govee_api_laggat import Govee, GoveeDevice, GoveeDeviceState
+from .govee_api_laggat import Govee, GoveeDevice
 from .__version__ import VERSION