+++ tests/test_govee_api_laggat.py
@@ -2,8 +2,9 @@ import asyncio
 from asynctest import TestCase, MagicMock, patch, CoroutineMock
 from aiohttp import ClientSession
 from datetime import datetime
+from time import time
 
-from govee_api_laggat import Govee, GoveeDevice, GoveeDeviceState
+from govee_api_laggat import Govee, GoveeDevice
 
 _API_URL = "https://developer-api.govee.com"
 _API_KEY = "SUPER_SECRET_KEY"
@@ -62,6 +63,13 @@ DUMMY_DEVICE_H6163 = GoveeDevice(
     support_brightness='brightness' in JSON_DEVICE_H6163['supportCmds'],
     support_color='color' in JSON_DEVICE_H6163['supportCmds'],
     support_color_tem='colorTem' in JSON_DEVICE_H6163['supportCmds'],
+    online=True,
+    power_state=True,
+    brightness=254,
+    color=(139, 0, 255),
+    timestamp = 0,
+    source = 'api', # this device supports status
+    error = None
 )
 DUMMY_DEVICE_H6104 = GoveeDevice(
     device=JSON_DEVICE_H6104['device'],
@@ -74,35 +82,17 @@ DUMMY_DEVICE_H6104 = GoveeDevice(
     support_brightness='brightness' in JSON_DEVICE_H6104['supportCmds'],
     support_color='color' in JSON_DEVICE_H6104['supportCmds'],
     support_color_tem='colorTem' in JSON_DEVICE_H6104['supportCmds'],
-)
-# light device state
-DUMMY_DEVICE_STATE_H6163 = GoveeDeviceState(
-    device=JSON_DEVICE_H6163['device'],
-    model=JSON_DEVICE_H6163['model'],
-    online=True,
-    power_state=True,
-    brightness=254,
-    color=(139, 0, 255),
-    timestamp = 0,
-    source = 'api' # this device supports status
-)
-DUMMY_DEVICE_STATE_H6104 = GoveeDeviceState(
-    device=JSON_DEVICE_H6104['device'],
-    model=JSON_DEVICE_H6104['model'],
     online=True,
     power_state=False,
     brightness=0,
     color=(0, 0, 0),
     timestamp = 0,
-    source = 'history'
+    source = 'history',
+    error = None
 )
 DUMMY_DEVICES = {
     DUMMY_DEVICE_H6163.device: DUMMY_DEVICE_H6163,
     DUMMY_DEVICE_H6104.device: DUMMY_DEVICE_H6104,
-}
-DUMMY_STATES = {
-    DUMMY_DEVICE_STATE_H6163.device: DUMMY_DEVICE_STATE_H6163,
-    DUMMY_DEVICE_STATE_H6104.device: DUMMY_DEVICE_STATE_H6104,
 }
 
 # json results for light states
@@ -147,7 +137,7 @@ class GoveeTests(TestCase):
 
         async def ping():
             async with Govee(_API_KEY) as govee:
-                return await govee.ping_async()
+                return await govee.ping()
         result, err = loop.run_until_complete(ping())
         # assert
         assert not err
@@ -161,8 +151,8 @@ class GoveeTests(TestCase):
         # arrange
         loop = asyncio.get_event_loop()
         mock_get.return_value.__aenter__.return_value.status = 200
-        mock_get.return_value.__aenter__.return_value.text = CoroutineMock(
-            return_value="Pong"
+        mock_get.return_value.__aenter__.return_value.json = CoroutineMock(
+            return_value=JSON_DEVICES
         )
         sleep_until = datetime.timestamp(datetime.now()) + 1
         mock_get.return_value.__aenter__.return_value.headers = {
@@ -172,22 +162,22 @@ class GoveeTests(TestCase):
         }
         mock_sleep.return_value.__aenter__.return_value.text = CoroutineMock()
         # act
-        async def ping():
+        async def get_devices():
             async with Govee(_API_KEY) as govee:
                 assert govee.rate_limit_on == 5
                 assert govee.rate_limit_total == 100
                 assert govee.rate_limit_reset == 0
                 assert govee.rate_limit_remaining == 100
                 # first run uses defaults, so ping returns immediatly
-                delay1, err1 = await govee.ping_async()
+                _, err1 = await govee.get_devices()
                 assert mock_sleep.call_count == 0
                 assert govee.rate_limit_remaining == 5
                 assert govee.rate_limit_reset == sleep_until
                 # second run, rate limit sleeps until the second is over
-                delay2, err2 = await govee.ping_async()
+                _, err2 = await govee.get_devices()
                 assert mock_sleep.call_count == 1
                 return err1, err2
-        err1, err2 = loop.run_until_complete(ping())
+        err1, err2 = loop.run_until_complete(get_devices())
         # assert
         assert not err1
         assert not err2
@@ -208,18 +198,17 @@ class GoveeTests(TestCase):
             _RATELIMIT_RESET: f'{sleep_until}'
         }
         # act
-
-        async def ping():
+        async def get_devices():
             async with Govee(_API_KEY) as govee:
                 assert govee.rate_limit_on == 5
                 assert govee.rate_limit_total == 100
                 assert govee.rate_limit_reset == 0
                 assert govee.rate_limit_remaining == 100
                 # first run uses defaults, so ping returns immediatly
-                return await govee.ping_async()
-        delay1, err1 = loop.run_until_complete(ping())
+                return await govee.get_devices()
+        result1, err1 = loop.run_until_complete(get_devices())
         # assert
-        assert not delay1
+        assert not result1
         assert err1 == 'API-Error 429: Rate limit exceeded, retry in 1 seconds.'
         assert mock_get.call_count == 1
 
@@ -228,8 +217,8 @@ class GoveeTests(TestCase):
         # arrange
         loop = asyncio.get_event_loop()
         mock_get.return_value.__aenter__.return_value.status = 200
-        mock_get.return_value.__aenter__.return_value.text = CoroutineMock(
-            return_value="Pong"
+        mock_get.return_value.__aenter__.return_value.json = CoroutineMock(
+            return_value=JSON_DEVICES
         )
         sleep_until = datetime.timestamp(datetime.now()) + 1
         mock_get.return_value.__aenter__.return_value.headers = {
@@ -238,20 +227,22 @@ class GoveeTests(TestCase):
             _RATELIMIT_RESET: f'{sleep_until}'
         }
         # act
-
-        async def ping():
+        async def get_devices():
             async with Govee(_API_KEY) as govee:
                 govee.rate_limit_on = 4
                 assert govee.rate_limit_on == 4  # set did work
                 # first run uses defaults, so ping returns immediatly
-                delay1, err1 = await govee.ping_async()
+                start = time()
+                _, err1 = await govee.get_devices()
+                delay1 = start - time()
                 # second run, doesn't rate limit either
-                delay2, err2 = await govee.ping_async()
+                _, err2 = await govee.get_devices()
+                delay2 = start - time()
                 return delay1, err1, delay2, err2
-        delay1, err1, delay2, err2 = loop.run_until_complete(ping())
+        delay1, err1, delay2, err2 = loop.run_until_complete(get_devices())
         # assert
-        assert delay1 < 10  # this should return immediatly
-        assert delay2 < 10  # this should return immediatly
+        assert delay1 < 0.10  # this should return immediatly
+        assert delay2 < 0.10  # this should return immediatly
         assert not err1
         assert not err2
         assert mock_get.call_count == 2
@@ -294,6 +285,7 @@ class GoveeTests(TestCase):
         async def getDevices():
             async with Govee(_API_KEY) as govee:
                 result, err = await govee.get_devices()
+                assert not err
                 cache = govee.devices
                 return result, cache
         result, cache = loop.run_until_complete(getDevices())
@@ -342,7 +334,6 @@ class GoveeTests(TestCase):
             async with Govee(_API_KEY) as govee:
                 # inject a device for testing
                 govee._devices = {DUMMY_DEVICE_H6163.device: DUMMY_DEVICE_H6163}
-                govee._states = {DUMMY_DEVICE_STATE_H6163.device: DUMMY_DEVICE_STATE_H6163}
                 return await govee.turn_on(DUMMY_DEVICE_H6163)
         success, err = loop.run_until_complete(turn_on())
         # assert
@@ -375,7 +366,6 @@ class GoveeTests(TestCase):
             async with Govee(_API_KEY) as govee:
                 # inject a device for testing
                 govee._devices = {DUMMY_DEVICE_H6163.device: DUMMY_DEVICE_H6163}
-                govee._states = {DUMMY_DEVICE_STATE_H6163.device: DUMMY_DEVICE_STATE_H6163}
                 return await govee.turn_on(DUMMY_DEVICE_H6163)
         success, err = loop.run_until_complete(turn_on())
         # assert
@@ -411,7 +401,6 @@ class GoveeTests(TestCase):
             async with Govee(_API_KEY) as govee:
                 # inject a device for testing
                 govee._devices = {DUMMY_DEVICE_H6163.device: DUMMY_DEVICE_H6163}
-                govee._states = {DUMMY_DEVICE_STATE_H6163.device: DUMMY_DEVICE_STATE_H6163}
                 # use device address here
                 return await govee.turn_off(DUMMY_DEVICE_H6163.device)
         success, err = loop.run_until_complete(turn_off())
@@ -433,7 +422,7 @@ class GoveeTests(TestCase):
 
     @patch('aiohttp.ClientSession.get')
     @patch('govee_api_laggat.Govee._state_request_allowed')
-    def test_get_state(self, mock_state_request_allowed, mock_get):
+    def test_get_states(self, mock_state_request_allowed, mock_get):
         # arrange
         loop = asyncio.get_event_loop()
         mock_state_request_allowed.return_value = True # always get live state
@@ -446,28 +435,24 @@ class GoveeTests(TestCase):
             async with Govee(_API_KEY) as govee:
                 # inject devices for testing
                 govee._devices = DUMMY_DEVICES
-                govee._states = DUMMY_STATES
-                results_per_device = {}
-                errors_per_device = {}
-                for dev in DUMMY_DEVICES:
-                    results_per_device[dev], errors_per_device[dev] = await govee.get_state(dev)
-                return results_per_device, errors_per_device
-        results, errors = loop.run_until_complete(getDevices())
+                #for dev in DUMMY_DEVICES:
+                #    results_per_device[dev], errors_per_device[dev] = await govee.get_state(dev)
+                states = await govee.get_states()
+                return states
+        states = loop.run_until_complete(getDevices())
         # assert
-        assert len(errors) == 2
-        assert errors[DUMMY_DEVICE_H6163.device] == None
-        assert errors[DUMMY_DEVICE_H6104.device] == None
         assert mock_get.call_count == 1 # only retrievable devices
         assert mock_get.call_args.kwargs['url'] == 'https://developer-api.govee.com/v1/devices/state'
         assert mock_get.call_args.kwargs['headers'] == {
-            'Govee-API-Key': 'SUPER_SECRET_KEY'}
-        assert mock_get.call_args.kwargs['params'] == {
-            'device': DUMMY_DEVICE_H6163.device, 'model': DUMMY_DEVICE_H6163.model}
-        assert len(results) == 2
-        assert isinstance(results[DUMMY_DEVICE_H6163.device], GoveeDeviceState)
-        DUMMY_DEVICE_STATE_H6163.timestamp = results[DUMMY_DEVICE_H6163.device].timestamp
-        assert results[DUMMY_DEVICE_H6163.device] == DUMMY_DEVICE_STATE_H6163
-        assert results[DUMMY_DEVICE_H6104.device] == DUMMY_DEVICE_STATE_H6104
+            'Govee-API-Key': 'SUPER_SECRET_KEY'
+        }
+        assert mock_get.call_args.kwargs['params']['device'] == DUMMY_DEVICE_H6163.device
+        assert mock_get.call_args.kwargs['params']['model'] == DUMMY_DEVICE_H6163.model
+        assert len(states) == 2
+        # to compare the 
+        DUMMY_DEVICE_H6163.timestamp = states[0].timestamp
+        assert states[0] == DUMMY_DEVICE_H6163
+        assert states[1] == DUMMY_DEVICE_H6104
         assert mock_state_request_allowed.call_count == 1
 
     @patch('aiohttp.ClientSession.put')
@@ -481,7 +466,6 @@ class GoveeTests(TestCase):
             async with Govee(_API_KEY) as govee:
                 # inject a device for testing
                 govee._devices = {DUMMY_DEVICE_H6163.device: DUMMY_DEVICE_H6163}
-                govee._states = {DUMMY_DEVICE_STATE_H6163.device: DUMMY_DEVICE_STATE_H6163}
                 return await govee.set_brightness(DUMMY_DEVICE_H6163, brightness)
         success, err = loop.run_until_complete(set_brightness())
         # assert
@@ -501,7 +485,6 @@ class GoveeTests(TestCase):
             async with Govee(_API_KEY) as govee:
                 # inject a device for testing
                 govee._devices = {DUMMY_DEVICE_H6163.device: DUMMY_DEVICE_H6163}
-                govee._states = {DUMMY_DEVICE_STATE_H6163.device: DUMMY_DEVICE_STATE_H6163}
                 return await govee.set_brightness(DUMMY_DEVICE_H6163, brightness)
         success, err = loop.run_until_complete(set_brightness())
         # assert
@@ -523,7 +506,6 @@ class GoveeTests(TestCase):
             async with Govee(_API_KEY) as govee:
                 # inject a device for testing
                 govee._devices = {DUMMY_DEVICE_H6163.device: DUMMY_DEVICE_H6163}
-                govee._states = {DUMMY_DEVICE_STATE_H6163.device: DUMMY_DEVICE_STATE_H6163}
                 return await govee.set_brightness(DUMMY_DEVICE_H6163.device, 42)
         success, err = loop.run_until_complete(set_brightness())
         # assert
@@ -556,7 +538,6 @@ class GoveeTests(TestCase):
             async with Govee(_API_KEY) as govee:
                 # inject a device for testing
                 govee._devices = {DUMMY_DEVICE_H6163.device: DUMMY_DEVICE_H6163}
-                govee._states = {DUMMY_DEVICE_STATE_H6163.device: DUMMY_DEVICE_STATE_H6163}
                 return await govee.set_color_temp(DUMMY_DEVICE_H6163.device, 6000)
         success, err = loop.run_until_complete(set_color_temp())
         # assert
@@ -589,7 +570,6 @@ class GoveeTests(TestCase):
             async with Govee(_API_KEY) as govee:
                 # inject a device for testing
                 govee._devices = {DUMMY_DEVICE_H6163.device: DUMMY_DEVICE_H6163}
-                govee._states = {DUMMY_DEVICE_STATE_H6163.device: DUMMY_DEVICE_STATE_H6163}
                 return await govee.set_color(DUMMY_DEVICE_H6163.device, (42, 43, 44))
         success, err = loop.run_until_complete(set_color())
         # assert
@@ -626,13 +606,11 @@ class GoveeTests(TestCase):
             async with Govee(_API_KEY) as govee:
                 # inject a device for testing
                 govee._devices = {DUMMY_DEVICE_H6163.device: DUMMY_DEVICE_H6163}
-                govee._states = {DUMMY_DEVICE_STATE_H6163.device: DUMMY_DEVICE_STATE_H6163}
                 await govee.turn_on(DUMMY_DEVICE_H6163)
                 # getting state to early (2s after switching)
-                return await govee.get_state(DUMMY_DEVICE_H6163)
-        state, err = loop.run_until_complete(turn_on_and_get_state())
+                return await govee.get_states()
+        states = loop.run_until_complete(turn_on_and_get_state())
         # assert
-        assert state.source == 'history'
-        assert err == None
+        assert states[0].source == 'history'
         assert mock_put.call_count == 1
         assert mock_get.call_count == 0 # may not get state